+++ pangea-sdk/pangea/pangea.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	version                = "5.3.0-beta.1"
+	version                = "5.3.0-beta.2"
 	pangeaUserAgent        = "pangea-go/" + version
 	serviceNamePlaceholder = "{SERVICE_NAME}"
 )